가수 선택하기 개발 진행 중 (로직 완료)

+++ core/firebase/src/main/java/com/manjee/firebase/model/RankingDataModel.kt
@@ -5,9 +5,9 @@ import com.google.firebase.database.IgnoreExtraProperties
 
 @IgnoreExtraProperties
 data class RankingDataModel(
-    val id: Int,
+    val id: Long,
     val name: String,
-    val score: Int
+    val score: Long
 ) {
 
     @Exclude
